抽出 CanvasImage
